backend/requirements.txt
# GroupChatLLM v3 Backend Dependencies

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.5
distro>=1.9.0

# AI Orchestration
langchain>=0.1.0
langgraph>=0.0.40
langsmith>=0.0.60
langchain-community>=0.0.10

# Model Providers
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
tiktoken>=0.9.0

# Async & Streaming
aiohttp>=3.9.0
sse-starlette>=1.6.0
httpx>=0.25.0

# Memory & Caching
redis>=4.5.0
aioredis>=2.0.0

# Data Processing
pydantic>=2.0.0
python-dateutil>=2.8.0
uuid6>=2022.10.25
pyyaml>=6.0.0

# Semantic Analysis
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Development & Testing
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Monitoring & Logging
loguru>=0.7.0
